Extract FilterButton from FilterComponent to be able to use them separately in the subheader
