+++ app/components/_index.sass
@@ -2,6 +2,6 @@
 @import "work_packages/share/invite_user_form_component"
 @import "work_packages/progress/modal_body_component"
 @import "open_project/common/attribute_component"
-@import "filters_component"
+@import "filter/filters_component"
 @import "projects/settings/project_custom_field_sections/index_component"
 @import "projects/row_component"
