+++ app/components/filter/filter_button_component.html.erb
@@ -1,0 +1,8 @@
+<%= render(Primer::Beta::Button.new(scheme: :secondary,
+                                    disabled:,
+                                    data: { "filters-target": "filterFormToggle",
+                                            action: "filters#toggleDisplayFilters" },
+                                    test_selector: "filter-component-toggle")) do |button| %>
+  <% button.with_trailing_visual_counter(count: filters_count) %>
+  <%= t(:label_filter) %>
+<% end %>
