+++ app/components/filter/filter_button_component.rb
@@ -27,60 +27,15 @@
 #
 # See COPYRIGHT and LICENSE files for more details.
 # ++
-module Meetings
+module Filter
   # rubocop:disable OpenProject/AddPreviewForViewComponent
-  class MeetingFiltersComponent < Filter::FilterComponent
+  class FilterButtonComponent < ApplicationComponent
     # rubocop:enable OpenProject/AddPreviewForViewComponent
-    options :project
-
-    def allowed_filters
-      super
-        .select { |f| allowed_filter?(f) }
-        .sort_by(&:human_name)
-    end
+    options :query
+    options :disabled
 
     def filters_count
-      @filters_count ||= begin
-        count = super
-        count -= 1 if project.present?
-
-        count
-      end
-    end
-
-    protected
-
-    def additional_filter_attributes(filter)
-      case filter
-      when Queries::Meetings::Filters::AuthorFilter,
-           Queries::Meetings::Filters::AttendedUserFilter,
-           Queries::Meetings::Filters::InvitedUserFilter
-        {
-          autocomplete_options: {
-            component: "opce-user-autocompleter",
-            resource: "principals"
-          }
-        }
-      else
-        super
-      end
-    end
-
-    private
-
-    def allowed_filter?(filter)
-      allowlist = [
-        Queries::Meetings::Filters::TimeFilter,
-        Queries::Meetings::Filters::AttendedUserFilter,
-        Queries::Meetings::Filters::InvitedUserFilter,
-        Queries::Meetings::Filters::AuthorFilter
-      ]
-
-      if project.nil?
-        allowlist << Queries::Meetings::Filters::ProjectFilter
-      end
-
-      allowlist.detect { |clazz| filter.is_a? clazz }
+      @filters_count ||= query.filters.count
     end
   end
 end
