+++ app/components/filter/filters_component.sass
@@ -43,10 +43,5 @@
     .advanced-filters--controls
       margin-top: 1rem
 
-  &-header
-    display: flex
-    justify-content: space-between
-    margin-bottom: 1rem
-    
-    &-actions
-      display: flex
+    &_top-margin
+      margin-top: 1rem
