+++ lookbook/previews/open_project/filter/filter_button_component_preview/filter_section_toggle.html.erb
@@ -1,0 +1,7 @@
+<div data-controller="filters"
+     data-application-target="dynamic"
+     data-filters-display-filters-value="false"
+     data-filters-output-format-value="params">
+  <%= render ::Filter::FilterButtonComponent.new(query: query) %>
+  <%= render Projects::ProjectsFiltersComponent.new(query: query) %>
+</div>
